Add native code for wrapping parts of libot
Using emscripten directly posed some issues with accessing struct fields
and dealing with offsets. Adding native functions that can return
specific fields of a struct instead of manually computing the offsets
makes the JS code a lot more maintainable.

There might be a way to tell emscripten to generate offset constants at
compile time (removing the need for native accessor functions), but I'm
not sure of how to do this.